inference/inference_benchmark/python/TensorFlow/requirements.txt
cup
pytz
py3nvml
tensorflow_gpu
